requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
astroid==2.11.5
asttokens==2.0.5
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
execnet==1.9.0
executing==0.8.3
glob2==0.7
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mysql-connector-python==8.0.30
nest-asyncio==1.5.5
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pylint==2.13.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-bdd==6.0.1
pytest-forked==1.4.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
pyzmq==23.1.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.3.0
tomli==2.0.1
tornado==6.1
traitlets==5.2.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
